♻️[Refactor] HeaderSection 공통 레이아웃 컴포넌트로 분리 #157
