+++ src/components/common-layout/index.ts
@@ -1,3 +1,5 @@
 export { default as FilterSection } from './FilterSection';
 export { default as SearchSection } from './SearchSection';
 export { default as CategoryTabs } from './CategoryTabs';
+export { default as FilterBar } from './FilterBar';
+export { default as HeaderSection } from './HeaderSection';
