+++ src/features/bookclub/components/BookClubMainPage.tsx
@@ -1,8 +1,9 @@
 'use client';
 
-import { HeaderSection, ClubListSection } from '@/features/bookclub/components';
+import { ClubListSection } from '@/features/bookclub/components';
 import useBookClubList from '../hooks/useFetchBookClubList';
-import FilterBar from '@/components/common-layout/FilterBar';
+import { FilterBar, HeaderSection } from '@/components/common-layout';
+import { useAuthStore } from '@/store/authStore';
 
 function BookClubMainPage() {
   // 커스텀 훅에서 상태와 핸들러 가져오기
@@ -12,9 +13,16 @@ function BookClubMainPage() {
     updateFilters(newFilter);
   };
 
+  const user = useAuthStore((state) => state.user);
+
+  const userName = user?.nickname || '북코';
+
   return (
     <>
-      <HeaderSection />
+      <HeaderSection showButton={true}>
+        반가워요, <span className="text-green-normal-01">{userName}</span>님!
+        <br />책 모임에 참여해 보세요
+      </HeaderSection>
       <FilterBar
         filters={filters}
         handleFilterChange={handleFilterChange}
